many to many products orders

+++ database/migrations/2017_05_11_115657_create_orders_table.php
@@ -17,7 +17,7 @@ class CreateOrdersTable extends Migration
             $table->increments('id');
             $table->date('buy_date');
             $table->date('pay_date')->nullable();
-            $table->decimal('total');
+            $table->double('total');
             $table->integer('client_id')->unsigned();
             $table->foreign('client_id')->references('id')->on('clients')->onDelete('cascade')->onUpdate('cascade');
         });
